backend/requirements.txt

fastapi
uvicorn
redis
pymupdf
markdownify
python-dotenv
litellm
python-multipart

